store class that handles carts/customer ids
